# backend/internal/api/router.go
package api

import (
	"net/http"
	"os"
	"path/filepath"
	"strings"
	"time"

	"dalu-nongji-parts/backend/internal/auth"
	"dalu-nongji-parts/backend/internal/config"
	"dalu-nongji-parts/backend/internal/model"
	"dalu-nongji-parts/backend/internal/service"
	"github.com/gin-contrib/cors"
	"github.com/gin-gonic/gin"
	"gorm.io/gorm"
)

type Deps struct {
	DB     *gorm.DB
	Config config.Config
}

func NewRouter(deps Deps) *gin.Engine {
	router := gin.New()
	accessProtection := NewAccessProtectionService(deps.DB, deps.Config)
	if err := router.SetTrustedProxies(deps.Config.TrustedProxyCIDRs); err != nil {
		panic("invalid trusted proxy configuration: " + err.Error())
	}
	router.Use(gin.Logger(), gin.Recovery())
	router.Use(func(c *gin.Context) {
		path := c.Request.URL.Path
		if strings.HasPrefix(path, "/assets/") {
			c.Header("Cache-Control", "public, max-age=31536000, immutable")
		} else if strings.HasPrefix(path, "/api/media/") || strings.HasPrefix(path, "/uploads/") || strings.HasPrefix(path, "/images/") {
			c.Header("Cache-Control", "public, max-age=86400, stale-while-revalidate=604800")
		} else if !strings.HasPrefix(path, "/api/") {
			c.Header("Cache-Control", "no-cache, no-store, must-revalidate")
		}
		c.Next()
	})
	router.Use(cors.New(cors.Config{
		AllowOrigins:     deps.Config.AllowedOrigins,
		AllowMethods:     []string{"GET", "POST", "PUT", "DELETE", "OPTIONS"},
		AllowHeaders:     []string{"Origin", "Content-Type", "Authorization", "X-CSRF-Token"},
		AllowCredentials: true,
	}))
	router.Use(accessProtection.Middleware())
	RegisterHealthRoute(router)
	RegisterPublicRoutesWithAuth(router, deps.DB, deps.Config.AuthSecret)
	RegisterMarketplaceRoutes(router, deps.DB, deps.Config)
	router.POST("/api/analytics/events", AnalyticsHandler{DB: deps.DB, Config: deps.Config}.RecordEvent)
	staticPages := NewStaticPageService(deps.DB, deps.Config)
	watermarks := NewWatermarkService(deps.DB, deps.Config, accessProtection)
	registerAdminRoutesWithServices(router, deps.DB, deps.Config, staticPages, accessProtection, watermarks)
	RegisterSEORoutes(router, deps.DB, deps.Config)
	mediaHandler := AdminHandler{DB: deps.DB, Config: deps.Config, Watermarks: watermarks}
	router.GET("/api/vendors/:id/contact-qr", CMSAuth(deps.DB, deps.Config.AuthSecret), mediaHandler.VendorContactQRCode)
	router.GET("/api/media/:id", mediaHandler.PublicMedia)
	registerStaticRoutesWithServices(router, deps.DB, deps.Config.PublicDir, staticPages, watermarks)
	return router
}

func RegisterHealthRoute(router *gin.Engine) {
	router.GET("/api/health", func(c *gin.Context) { OK(c, gin.H{"status": "ok"}) })
}

func RegisterPublicRoutes(router *gin.Engine, db *gorm.DB) {
	RegisterPublicRoutesWithAuth(router, db, "")
}

func RegisterPublicRoutesWithAuth(router *gin.Engine, db *gorm.DB, secret string) {
	handler := PublicHandler{DB: db, HomeService: service.HomeService{DB: db}}
	api := router.Group("/api")
	api.Use(PublicETag())
	api.Use(OptionalCMSAuth(db, secret))
	api.GET("/home", handler.Home)
	api.GET("/site-meta", handler.SiteMeta)
	api.GET("/layout-config", handler.LayoutConfig)
	api.GET("/menus", handler.Menus)
	api.GET("/pages/:slug", handler.Page)
	api.GET("/articles", handler.Articles)
	api.GET("/articles/:slug", handler.Article)
	api.GET("/friend-links", handler.FriendLinks)
	api.GET("/processing-vendors", handler.ProcessingVendors)
	api.GET("/processing-filter-options", handler.ProcessingFilterOptions)
	api.GET("/vendors", handler.Vendors)
	api.GET("/vendor-categories", handler.VendorCategories)
	api.GET("/vendors/recommended", handler.RecommendedVendors)
	api.GET("/vendors/slug/:slug", handler.VendorBySlug)
	api.GET("/vendors/:id/contact", CMSAuth(db, secret), handler.VendorContact)
	api.GET("/vendors/:id/posts", handler.VendorPosts)
	api.GET("/vendors/:id", handler.VendorDetail)
	api.GET("/products", handler.Products)
	api.GET("/products/slug/:slug", handler.ProductBySlug)
	api.GET("/products/:id", handler.ProductDetail)
	api.GET("/products/:id/suppliers", handler.ProductSuppliers)
	api.GET("/search", handler.Search)
	api.GET("/filter-options", handler.FilterOptions)
	api.GET("/categories/slug/:slug", handler.CategoryBySlug)
}

// OptionalCMSAuth enriches public requests when a valid CMS account token is
// present. Invalid or missing tokens remain anonymous and never block browsing.
func OptionalCMSAuth(db *gorm.DB, secret string) gin.HandlerFunc {
	return func(c *gin.Context) {
		if user, ok := authenticateRequest(c, db, secret); ok {
			c.Set("authenticated", true)
			setUserContext(c, user)
		}
		c.Next()
	}
}

func RegisterAdminRoutes(router *gin.Engine, db *gorm.DB, cfg config.Config) {
	accessProtection := NewAccessProtectionService(db, cfg)
	registerAdminRoutesWithServices(router, db, cfg, NewStaticPageService(db, cfg), accessProtection, NewWatermarkService(db, cfg, accessProtection))
}

// RegisterMarketplaceRoutes contains only additive v1 endpoints. Existing web
// and CMS routes remain unchanged and can be deployed before either new client.
func RegisterMarketplaceRoutes(router *gin.Engine, db *gorm.DB, cfg config.Config) {
	handler := AdminHandler{DB: db, Config: cfg}
	appAuth := router.Group("/api/v1/app/auth")
	appAuth.POST("/register", handler.AppRegister)
	appAuth.POST("/login", handler.AppLogin)
	appAuth.POST("/refresh", handler.AppRefresh)
	appProtected := appAuth.Group("")
	appProtected.Use(CMSAuth(db, cfg.AuthSecret))
	appProtected.POST("/logout", handler.AppLogout)
	appProtected.GET("/me", handler.AppMe)

	market := router.Group("/api/v1/market-posts")
	market.GET("", handler.ListMarketPosts)
	market.GET("/:id", handler.MarketPostDetail)
	marketContact := market.Group("")
	marketContact.Use(CMSAuth(db, cfg.AuthSecret))
	marketContact.GET("/:id/contact", handler.MarketPostContact)

	auctions := router.Group("/api/v1/auctions")
	auctions.Use(OptionalCMSAuth(db, cfg.AuthSecret))
	auctions.GET("", handler.ListPublicAuctions)
	auctions.GET("/:id", handler.PublicAuctionDetail)
	auctionBid := auctions.Group("")
	auctionBid.Use(CMSAuth(db, cfg.AuthSecret), RequireCSRF(), RequireRole("vendor"))
	auctionBid.POST("/:id/bids", handler.PlaceAuctionBid)

	account := router.Group("/api/v1/me")
	account.Use(CMSAuth(db, cfg.AuthSecret), RequireCSRF(), RequireAnyRole("buyer", "vendor"))
	account.GET("/market-posts", handler.ListOwnMarketPosts)
	account.GET("/market-posts/:id", handler.OwnMarketPost)
	account.POST("/market-posts", handler.CreateOwnMarketPost)
	account.PUT("/market-posts/:id", handler.UpdateOwnMarketPost)
	account.DELETE("/market-posts/:id", handler.WithdrawOwnMarketPost)
	account.GET("/profile", handler.AccountProfile)
	account.PUT("/profile", handler.UpdateAccountProfile)
	account.GET("/auctions", handler.ListOwnAuctions)
	account.GET("/auctions/:id", handler.OwnAuctionDetail)
	account.POST("/auctions", handler.CreateOwnAuction)
	account.PUT("/auctions/:id", handler.UpdateOwnAuction)
	account.POST("/auctions/:id/publish", handler.PublishOwnAuction)
	account.POST("/auctions/:id/cancel", handler.CancelOwnAuction)
	account.POST("/auctions/:id/award", handler.AwardOwnAuction)
	account.GET("/notifications", handler.ListNotifications)
	account.PUT("/notifications/read-all", handler.ReadAllNotifications)
	account.PUT("/notifications/:id/read", handler.ReadNotification)

	media := router.Group("/api/v1")
	media.Use(CMSAuth(db, cfg.AuthSecret), RequireCSRF(), RequireAnyRole("buyer", "vendor"))
	media.POST("/media", handler.SecureUpload)
}

func registerAdminRoutesWithServices(router *gin.Engine, db *gorm.DB, cfg config.Config, staticPages *StaticPageService, accessProtection *AccessProtectionService, watermarks *WatermarkService) {
	handler := AdminHandler{DB: db, Config: cfg, StaticPages: staticPages, AccessProtection: accessProtection, Watermarks: watermarks}
	authAPI := router.Group("/api/auth")
	authAPI.POST("/login", handler.Login)
	authAPI.POST("/register", handler.Register)
	authProtected := authAPI.Group("")
	authProtected.Use(CMSAuth(db, cfg.AuthSecret), RequireCSRF())
	authProtected.GET("/session", handler.CurrentSession)
	authProtected.POST("/logout", handler.Logout)
	authProtected.PUT("/password", handler.ChangePassword)

	admin := router.Group("/api/admin")
	admin.POST("/login", handler.StaffLogin)
	protected := admin.Group("")
	protected.Use(CMSAuth(db, cfg.AuthSecret))
	protected.Use(RequireCSRF())
	protected.GET("/profile", handler.Profile)

	workflowReaders := protected.Group("")
	workflowReaders.Use(RequireAnyRole("admin", "editor", "reviewer"))
	workflowReaders.GET("/dashboard", handler.DashboardStats)
	workflowReaders.GET("/revisions", handler.ListRevisions)
	workflowReaders.GET("/revisions/:id/preview", handler.PreviewRevision)
	workflowReaders.GET("/editorial/vendors", handler.ListVendors)
	workflowReaders.GET("/editorial/vendor-options", handler.ListVendorOptions)
	workflowReaders.GET("/editorial/products", handler.ListProducts)
	workflowReaders.GET("/editorial/categories", handler.ListCategories)
	workflowReaders.GET("/editorial/vendor-categories", handler.ListVendorCategories)
	workflowReaders.GET("/editorial/pages", handler.ListPages)
	workflowReaders.GET("/editorial/tags", handler.ListTags)

	workflowEditors := protected.Group("")
	workflowEditors.Use(RequireAnyRole("admin", "editor"))
	workflowEditors.POST("/seo-suggestions/vendor", handler.SuggestVendorSEO)
	workflowEditors.POST("/revisions", handler.SaveRevision)
	workflowEditors.POST("/revisions/:id/submit", handler.SubmitRevision)
	workflowEditors.GET("/vendor-categories", handler.ListVendorCategories)
	workflowEditors.POST("/vendor-categories", handler.CreateVendorCategory)
	workflowEditors.PUT("/vendor-categories/:id", handler.UpdateVendorCategory)
	workflowEditors.DELETE("/vendor-categories/:id", handler.DeleteVendorCategory)

	workflowReviewers := protected.Group("")
	workflowReviewers.Use(RequireAnyRole("admin", "reviewer"))
	workflowReviewers.POST("/revisions/:id/approve", handler.ApproveRevision)
	workflowReviewers.POST("/revisions/:id/reject", handler.RejectRevision)
	workflowReviewers.POST("/revisions/:id/archive", handler.ArchiveRevisionResource)
	workflowReviewers.GET("/market-posts", handler.AdminListMarketPosts)
	workflowReviewers.PUT("/market-posts/:id/status", handler.AdminUpdateMarketPostStatus)
	workflowReviewers.PUT("/vendor-posts/:id/review", handler.ReviewVendorPost)
	workflowReviewers.GET("/auctions", handler.AdminListAuctions)
	workflowReviewers.GET("/auctions/:id", handler.AdminAuctionDetail)
	workflowReviewers.PUT("/auctions/:id/status", handler.AdminUpdateAuctionStatus)

	mediaUsers := protected.Group("")
	mediaUsers.Use(RequireAnyRole("admin", "editor", "reviewer", "vendor"))
	mediaUsers.POST("/uploads", handler.SecureUpload)
	mediaUsers.POST("/remote-images", handler.DownloadRemoteImage)
	mediaUsers.GET("/media/:id", handler.PreviewMedia)
	mediaUsers.PUT("/media/:id/metadata", handler.UpdateMediaMetadata)
	mediaUsers.DELETE("/media/:id", handler.DeleteMedia)

	cmsOnly := protected.Group("")
	cmsOnly.Use(RequireAnyRole("admin", "vendor"))
	cmsOnly.GET("/vendor-profile", handler.GetVendorProfile)
	cmsOnly.PUT("/vendor-profile", handler.SubmitVendorProfile)
	cmsOnly.GET("/vendor-products", handler.ListOwnProducts)
	cmsOnly.GET("/vendor-products/duplicate-check", handler.CheckOwnProductDuplicate)
	cmsOnly.POST("/vendor-products", handler.CreateOwnProduct)
	cmsOnly.PUT("/vendor-product-submissions/:id", handler.UpdateOwnProductSubmission)
	cmsOnly.DELETE("/vendor-product-submissions/:id", handler.WithdrawOwnProductSubmission)
	cmsOnly.PUT("/vendor-products/:id", handler.UpdateOwnProduct)
	cmsOnly.DELETE("/vendor-products/:id", handler.DeleteOwnProduct)
	cmsOnly.PUT("/vendor-products/:id/price", handler.UpdateOwnProductPrice)
	cmsOnly.GET("/vendor-products/:id/price-history", handler.OwnProductPriceHistory)

	vendorOnly := protected.Group("")
	vendorOnly.Use(RequireRole("vendor"))
	vendorOnly.GET("/vendor-analytics", AnalyticsHandler{DB: db, Config: cfg}.VendorSummary)
	vendorOnly.POST("/vendor-posts", handler.CreateOwnVendorPost)
	vendorOnly.PUT("/vendor-posts/:id", handler.UpdateOwnVendorPost)
	vendorOnly.POST("/vendor-posts/:id/submit", handler.SubmitOwnVendorPost)
	vendorOnly.POST("/vendor-posts/:id/withdraw", handler.WithdrawOwnVendorPost)

	vendorPostReaders := protected.Group("")
	vendorPostReaders.Use(RequireAnyRole("admin", "reviewer", "vendor"))
	vendorPostReaders.GET("/vendor-posts", handler.ListVendorPostsForRole)

	adminOnly := protected.Group("")
	adminOnly.Use(RequireRole("admin"))
	adminOnly.GET("/analytics", AnalyticsHandler{DB: db, Config: cfg}.Summary)
	adminOnly.GET("/seo-status", handler.SEOConsoleHealth)
	adminOnly.POST("/seo-submit/baidu", handler.SubmitBaiduURLs)
	adminOnly.GET("/operation-logs", handler.ListOperationLogs)
	adminOnly.POST("/imports/:resource", handler.ImportWorkbook)
	adminOnly.GET("/menus", handler.ListMenus)
	adminOnly.POST("/menus", handler.CreateMenu)
	adminOnly.PUT("/menus/:id", handler.UpdateMenu)
	adminOnly.DELETE("/menus/:id", handler.DeleteMenu)
	adminOnly.GET("/vendors", handler.ListVendors)
	adminOnly.GET("/vendor-options", handler.ListVendorOptions)
	adminOnly.POST("/vendors", handler.CreateVendor)
	adminOnly.PUT("/vendors/:id", handler.UpdateVendor)
	adminOnly.DELETE("/vendors/:id", handler.DeleteVendor)
	adminOnly.GET("/vendor-submissions", handler.ListVendorSubmissions)
	adminOnly.PUT("/vendor-submissions/:id/review", handler.ReviewVendorSubmission)
	adminOnly.GET("/product-submissions", handler.ListProductSubmissions)
	adminOnly.GET("/product-submissions/:id/matches", handler.ProductSubmissionMatches)
	adminOnly.PUT("/product-submissions/:id/review", handler.ReviewProductSubmission)
	adminOnly.GET("/users", handler.ListCMSUsers)
	adminOnly.POST("/users", handler.CreateCMSUser)
	adminOnly.PUT("/users/:id", handler.UpdateCMSUser)
	adminOnly.DELETE("/users/:id", handler.DeleteCMSUser)
	adminOnly.GET("/tags", handler.ListTags)
	adminOnly.POST("/tags", handler.CreateTag)
	adminOnly.PUT("/tags/:id", handler.UpdateTag)
	adminOnly.DELETE("/tags/:id", handler.DeleteTag)
	adminOnly.GET("/categories", handler.ListCategories)
	adminOnly.POST("/categories", handler.CreateCategory)
	adminOnly.PUT("/categories/:id", handler.UpdateCategory)
	adminOnly.DELETE("/categories/:id", handler.DeleteCategory)
	adminOnly.GET("/products", handler.ListProducts)
	adminOnly.POST("/products", handler.CreateProduct)
	adminOnly.PUT("/products/:id", handler.UpdateProduct)
	adminOnly.DELETE("/products/:id", handler.DeleteProduct)
	adminOnly.GET("/products/:id/suppliers", handler.ListAdminProductSuppliers)
	adminOnly.POST("/products/:id/suppliers", handler.SaveAdminProductSupplier)
	adminOnly.DELETE("/products/:id/suppliers/:supplierId", handler.DisableAdminProductSupplier)
	adminOnly.POST("/product-suppliers/batch", handler.BatchSaveAdminProductSuppliers)
	adminOnly.PUT("/products/:id/merge", handler.MergeProducts)
	adminOnly.GET("/banners", handler.ListBanners)
	adminOnly.POST("/banners", handler.CreateBanner)
	adminOnly.PUT("/banners/:id", handler.UpdateBanner)
	adminOnly.DELETE("/banners/:id", handler.DeleteBanner)
	adminOnly.GET("/pages", handler.ListPages)
	adminOnly.POST("/pages", handler.CreatePage)
	adminOnly.PUT("/pages/:id", handler.UpdatePage)
	adminOnly.DELETE("/pages/:id", handler.DeletePage)
	adminOnly.GET("/friend-links", handler.ListFriendLinks)
	adminOnly.POST("/friend-links", handler.CreateFriendLink)
	adminOnly.PUT("/friend-links/:id", handler.UpdateFriendLink)
	adminOnly.DELETE("/friend-links/:id", handler.DeleteFriendLink)
	adminOnly.GET("/configs", handler.ListConfigs)
	adminOnly.PUT("/configs/:key", handler.UpdateConfig)
	adminOnly.GET("/static-pages/status", handler.StaticPageSummary)
	adminOnly.PUT("/static-pages/settings", handler.UpdateStaticPageSettings)
	adminOnly.GET("/static-pages/resources", handler.StaticPageResourceStatuses)
	adminOnly.POST("/static-pages/jobs", handler.CreateStaticBuildJob)
	adminOnly.GET("/static-pages/jobs/:id", handler.StaticBuildJob)
	adminOnly.GET("/access-protection/status", handler.ProtectionStatus)
	adminOnly.PUT("/access-protection/config", handler.UpdateProtectionConfig)
	adminOnly.GET("/access-protection/events", handler.ProtectionEvents)
	adminOnly.GET("/access-protection/blocks", handler.ProtectionBlocks)
	adminOnly.POST("/access-protection/blocks", handler.CreateProtectionBlock)
	adminOnly.POST("/access-protection/blocks/:id/release", handler.ReleaseProtectionBlock)
	adminOnly.POST("/access-protection/watermarks/jobs", handler.CreateWatermarkBuildJob)
	adminOnly.GET("/access-protection/watermarks/jobs/:id", handler.WatermarkBuildJob)
}

func AdminAuth(secret string) gin.HandlerFunc {
	return func(c *gin.Context) {
		header := c.GetHeader("Authorization")
		if !strings.HasPrefix(header, "Bearer ") {
			Fail(c, 401, 401, "未登录或登录已过期")
			c.Abort()
			return
		}
		username, err := auth.ParseToken(strings.TrimPrefix(header, "Bearer "), secret)
		if err != nil {
			Fail(c, 401, 401, "登录已过期，请重新登录")
			c.Abort()
			return
		}
		c.Set("username", username)
		c.Next()
	}
}

// CMSAuth enriches a valid token with the current user's role and vendor scope.
// Keeping AdminAuth separate preserves the small stateless middleware used by tests.
func CMSAuth(db *gorm.DB, secret string) gin.HandlerFunc {
	return func(c *gin.Context) {
		if db != nil {
			user, ok := authenticateRequest(c, db, secret)
			if !ok {
				Fail(c, http.StatusUnauthorized, 401, "未登录、会话已过期或账号已停用")
				c.Abort()
				return
			}
			setUserContext(c, user)
			c.Next()
			return
		}
		header := c.GetHeader("Authorization")
		if !strings.HasPrefix(header, "Bearer ") {
			Fail(c, http.StatusUnauthorized, 401, "未登录或登录已过期")
			c.Abort()
			return
		}
		username, err := auth.ParseToken(strings.TrimPrefix(header, "Bearer "), secret)
		if err != nil {
			Fail(c, http.StatusUnauthorized, 401, "登录已过期，请重新登录")
			c.Abort()
			return
		}
		c.Set("username", username)
		if db == nil {
			c.Set("role", "admin")
			c.Next()
			return
		}
		var user model.AdminUser
		if err := db.Where("username = ? AND is_enabled = ?", c.GetString("username"), true).First(&user).Error; err != nil {
			Fail(c, http.StatusUnauthorized, 401, "账号不存在或已停用")
			c.Abort()
			return
		}
		role := strings.TrimSpace(user.Role)
		if role == "" {
			role = "admin"
		}
		c.Set("role", role)
		if user.VendorID != nil {
			c.Set("vendorId", *user.VendorID)
		}
		c.Next()
	}
}

func setUserContext(c *gin.Context, user model.AdminUser) {
	role := strings.TrimSpace(user.Role)
	if role == "" {
		role = "admin"
	}
	c.Set("username", user.Username)
	c.Set("role", role)
	c.Set("userId", user.ID)
	if user.VendorID != nil {
		c.Set("vendorId", *user.VendorID)
	}
}

func RequireRole(role string) gin.HandlerFunc {
	return func(c *gin.Context) {
		if c.GetString("role") != role {
			Fail(c, http.StatusForbidden, 403, "没有权限执行此操作")
			c.Abort()
			return
		}
		c.Next()
	}
}

func RequireAnyRole(roles ...string) gin.HandlerFunc {
	allowed := make(map[string]struct{}, len(roles))
	for _, role := range roles {
		allowed[role] = struct{}{}
	}
	return func(c *gin.Context) {
		if _, ok := allowed[c.GetString("role")]; !ok {
			Fail(c, http.StatusForbidden, 403, "没有权限执行此操作")
			c.Abort()
			return
		}
		c.Next()
	}
}

func RegisterStaticRoutes(router *gin.Engine, db *gorm.DB, publicDir string, staticPageServices ...*StaticPageService) {
	var staticPages *StaticPageService
	if len(staticPageServices) > 0 {
		staticPages = staticPageServices[0]
	}
	registerStaticRoutesWithServices(router, db, publicDir, staticPages, nil)
}

func registerStaticRoutesWithServices(router *gin.Engine, db *gorm.DB, publicDir string, staticPages *StaticPageService, watermarks *WatermarkService) {
	uploadsDir := "uploads"
	if publicDir != "" {
		uploadsDir = filepath.Join(publicDir, "uploads")
	}
	if _, err := os.Stat(uploadsDir); err == nil {
		router.GET("/uploads/:name", func(c *gin.Context) {
			name := filepath.Base(c.Param("name"))
			if name == "." || name == "" || name != c.Param("name") {
				c.Status(http.StatusNotFound)
				return
			}
			url := "/uploads/" + name
			var count int64
			db.Model(&model.Vendor{}).Where("is_visible = ? AND publication_status = ? AND (published_at IS NULL OR published_at <= ?) AND (logo = ? OR cover_image = ?)", true, "published", time.Now(), url, url).Count(&count)
			if count == 0 {
				db.Model(&model.VendorMedia{}).Joins("JOIN vendors ON vendors.id = vendor_media.vendor_id").Where("vendor_media.url = ? AND vendors.is_visible = ? AND vendors.publication_status = ?", url, true, "published").Count(&count)
			}
			if count == 0 {
				db.Model(&model.Product{}).Where("products.publication_status = ? AND (products.image = ? OR products.gallery LIKE ? OR products.specs LIKE ?) AND EXISTS (SELECT 1 FROM product_suppliers ps JOIN vendors v ON v.id = ps.vendor_id WHERE ps.product_id = products.id AND ps.status = 'approved' AND v.is_visible = 1 AND v.publication_status = 'published')", "published", url, "%"+url+"%", "%"+url+"%").Count(&count)
				if count == 0 {
					db.Model(&model.ProductSupplier{}).Joins("JOIN vendors ON vendors.id = product_suppliers.vendor_id").Where("product_suppliers.status = ? AND vendors.is_visible = ? AND vendors.publication_status = ? AND (product_suppliers.image = ? OR product_suppliers.gallery LIKE ? OR product_suppliers.specs LIKE ?)", "approved", true, "published", url, "%"+url+"%", "%"+url+"%").Count(&count)
				}
			}
			if count == 0 {
				db.Model(&model.Banner{}).Where("is_enabled = ? AND background_image = ?", true, url).Count(&count)
			}
			if count == 0 {
				c.Status(http.StatusNotFound)
				return
			}
			publicPath := filepath.Join(uploadsDir, name)
			if watermarks != nil {
				var publicMIME string
				var err error
				publicPath, publicMIME, err = watermarks.PublicLegacyPath(publicPath, url)
				if err != nil {
					Fail(c, http.StatusServiceUnavailable, 503, "公开图片生成失败，请稍后重试")
					return
				}
				if publicMIME != "" {
					c.Header("Content-Type", publicMIME)
				}
			}
			c.File(publicPath)
		})
	}
	publicDirReady := false
	if publicDir != "" {
		if _, err := os.Stat(publicDir); err == nil {
			publicDirReady = true
			assetsDir := filepath.Join(publicDir, "assets")
			if _, err := os.Stat(assetsDir); err == nil {
				router.Static("/assets", assetsDir)
			}
			imagesDir := filepath.Join(publicDir, "images")
			if _, err := os.Stat(imagesDir); err == nil {
				router.Static("/images", imagesDir)
			}
		}
	}
	router.NoRoute(func(c *gin.Context) {
		if strings.HasPrefix(c.Request.URL.Path, "/api/") {
			Fail(c, http.StatusNotFound, 404, "接口不存在")
			return
		}
		if publicDirReady && staticPages != nil && staticPages.Serve(c) {
			return
		}
		if RenderSEOApp(c, db, config.Config{PublicDir: publicDir}, publicDir) {
			return
		}
		if publicDirReady {
			c.File(filepath.Join(publicDir, "index.html"))
			return
		}
		c.Status(http.StatusNotFound)
	})
}