# shared/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 大陆农机配件公共 REST API
  version: 1.0.0
  description: 新增移动端与供求接口的语言无关契约；既有 /api 接口保持兼容。
servers:
  - url: /
paths:
  /api/v1/app/auth/register:
    post:
      operationId: appRegister
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppRegistration' }
      responses:
        '200': { $ref: '#/components/responses/AppTokens' }
  /api/v1/app/auth/login:
    post:
      operationId: appLogin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AppLogin' }
      responses:
        '200': { $ref: '#/components/responses/AppTokens' }
  /api/v1/app/auth/refresh:
    post:
      operationId: appRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties: { refreshToken: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/AppTokens' }
  /api/v1/app/auth/logout:
    post:
      operationId: appLogout
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/app/auth/me:
    get:
      operationId: appMe
      security: [{ bearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/market-posts:
    get:
      operationId: listMarketPosts
      parameters:
        - { name: type, in: query, schema: { $ref: '#/components/schemas/MarketPostType' } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: categoryId, in: query, schema: { type: integer } }
        - { name: province, in: query, schema: { type: string } }
        - { name: publishedAfter, in: query, schema: { type: string, format: date-time } }
        - { name: publishedBefore, in: query, schema: { type: string, format: date-time } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: 公开供求列表，不含联系方式
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/MarketPostPage' }
  /api/v1/market-posts/{id}:
    get:
      operationId: getMarketPost
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: 公开供求详情，不含联系方式
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/MarketPost' } }
  /api/v1/market-posts/{id}/contact:
    get:
      operationId: getMarketPostContact
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200':
          description: 登录并通过访问限频后返回
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/ApiResponse' }
                  - type: object
                    properties: { data: { $ref: '#/components/schemas/MarketPostContact' } }
  /api/v1/me/market-posts:
    get:
      operationId: listOwnMarketPosts
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      operationId: createMarketPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketPostInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/market-posts/{id}:
    get:
      operationId: getOwnMarketPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      operationId: updateOwnMarketPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MarketPostInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
    delete:
      operationId: withdrawOwnMarketPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/profile:
    get:
      operationId: getAccountProfile
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      operationId: updateBuyerProfile
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuyerProfileInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/media:
    post:
      operationId: uploadMarketMedia
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/admin/market-posts:
    get:
      operationId: adminListMarketPosts
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/MarketPostStatus' } }
        - { name: keyword, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, minimum: 1, maximum: 100 } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/admin/market-posts/{id}/status:
    put:
      operationId: adminUpdateMarketPostStatus
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [published, removed] }
                reason: { type: string, maxLength: 500 }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/vendors/{id}/posts:
    get:
      operationId: listPublicVendorPosts
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/admin/vendor-products/{id}/price:
    put:
      operationId: updateVendorProductPrice
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/SupplierPriceInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { description: 价格版本冲突 }
  /api/admin/vendor-posts:
    get:
      operationId: listVendorPostsForRole
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      operationId: createVendorPost
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/VendorPostInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auctions:
    get:
      operationId: listAuctions
      parameters:
        - { name: keyword, in: query, schema: { type: string } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auctions/{id}:
    get:
      operationId: getAuction
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/auctions/{id}/bids:
    post:
      operationId: placeAuctionBid
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuctionBidInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { description: 竞价版本冲突、已截止或报价未降低 }
  /api/v1/me/auctions:
    get:
      operationId: listOwnAuctions
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      operationId: createAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuctionInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/auctions/{id}:
    get:
      operationId: getOwnAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
    put:
      operationId: updateAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AuctionInput' } } }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '409': { description: 竞价版本冲突 }
  /api/v1/me/auctions/{id}/publish:
    post:
      operationId: publishAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/auctions/{id}/cancel:
    post:
      operationId: cancelAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/auctions/{id}/award:
    post:
      operationId: awardAuction
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Id' }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
  /api/v1/me/notifications:
    get:
      operationId: listNotifications
      security: [{ bearerAuth: [] }, { cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
    cookieAuth: { type: apiKey, in: cookie, name: dlc_session }
  parameters:
    Id: { name: id, in: path, required: true, schema: { type: integer, minimum: 1 } }
  responses:
    Success:
      description: 成功
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiResponse' } } }
    AppTokens:
      description: App 登录或刷新成功
      content:
        application/json:
          schema:
            allOf:
              - { $ref: '#/components/schemas/ApiResponse' }
              - type: object
                properties: { data: { $ref: '#/components/schemas/AppSessionTokens' } }
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer }
        message: { type: string }
        data: {}
...
    MarketPostType:
      type: string
      enum: [supply, demand]
...
    AppLogin:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
        deviceName: { type: string, maxLength: 120 }
    AppRegistration:
      allOf:
        - { $ref: '#/components/schemas/AppLogin' }
        - type: object
          properties:
            role: { type: string, enum: [buyer, vendor], default: buyer }
            companyName: { type: string }
            displayName: { type: string }
            contactName: { type: string }
            phone: { type: string }
            province: { type: string }
            city: { type: string }
...